eatgo-customer-api/src/main/java/com/alexcode/eatgo/interfaces/RegionController.java
package com.alexcode.eatgo.interfaces;

import com.alexcode.eatgo.application.RegionService;
import com.alexcode.eatgo.domain.network.SuccessResponse;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.security.access.prepost.PreAuthorize;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

@RestController
@RequestMapping(path = "customer/api/v1/regions")
public class RegionController {

  @Autowired
  private RegionService regionService;

  @GetMapping(produces = MediaType.APPLICATION_JSON_VALUE)
  @PreAuthorize("hasAuthority('region:read')")
  public ResponseEntity<SuccessResponse> list() {
    SuccessResponse body = regionService.list();
    return new ResponseEntity<>(body, HttpStatus.OK);
  }

}
